Split morale string functions out of `Common.h`
